Add more structure to strat engine
Isolate parts of thinpool into a subdirectory.
They were already naturally isolated, as can be seen from the small number
of values that are exported from the thinpool subdirectory.
