+++ src/engine/strat_engine/thinpool/mod.rs
@@ -2,20 +2,11 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-mod blockdev;
-mod blockdevmgr;
-mod cleanup;
-mod device;
-mod engine;
-mod metadata;
-mod pool;
-mod serde_structs;
-mod setup;
-mod range_alloc;
+mod dmdevice;
+mod filesystem;
+mod mdv;
+#[allow(module_inception)]
 mod thinpool;
 mod util;
 
-pub use self::engine::StratEngine;
-
-#[cfg(test)]
-mod tests;
+pub use self::thinpool::{DATA_BLOCK_SIZE, DATA_LOWATER, INITIAL_DATA_SIZE, ThinPool};
