+++ src/engine/strat_engine/thinpool/thinpool.rs
@@ -14,20 +14,20 @@ use devicemapper::{DM, DataBlocks, DmDevice, DmName, DmNameBuf, IEC, LinearDev, 
                    Sectors, Segment, ThinDev, ThinDevId, ThinPoolDev, ThinPoolStatusSummary,
                    device_exists};
 
-use super::super::engine::{Filesystem, HasName};
-use super::super::errors::{EngineError, EngineResult, ErrorEnum};
-use super::super::structures::Table;
-use super::super::types::{DevUuid, PoolUuid, FilesystemUuid, RenameAction};
+use super::super::super::engine::{Filesystem, HasName};
+use super::super::super::errors::{EngineError, EngineResult, ErrorEnum};
+use super::super::super::structures::Table;
+use super::super::super::types::{DevUuid, PoolUuid, FilesystemUuid, RenameAction};
 
-use super::blockdevmgr::{BlockDevMgr, BlkDevSegment, map_to_dm};
-use super::device::wipe_sectors;
+use super::super::blockdevmgr::{BlockDevMgr, BlkDevSegment, map_to_dm};
+use super::super::device::wipe_sectors;
+use super::super::serde_structs::{FilesystemSave, FlexDevsSave, Recordable, ThinPoolDevSave};
+
 use super::dmdevice::{FlexRole, ThinDevIdPool, ThinPoolRole, ThinRole, format_flex_name,
                       format_thinpool_name, format_thin_name};
 use super::filesystem::{FilesystemStatus, StratFilesystem};
 use super::mdv::MetadataVol;
-use super::serde_structs::{FilesystemSave, FlexDevsSave, Recordable, ThinPoolDevSave};
 use super::util::execute_cmd;
-
 
 pub const DATA_BLOCK_SIZE: Sectors = Sectors(2048);
 pub const DATA_LOWATER: DataBlocks = DataBlocks(512);
@@ -641,10 +641,11 @@ mod tests {
 
     use devicemapper::{Bytes, SECTOR_SIZE};
 
+    use super::super::super::metadata::MIN_MDA_SECTORS;
+    use super::super::super::tests::{loopbacked, real};
+    use super::super::super::tests::tempdir::TempDir;
+
     use super::super::filesystem::{FILESYSTEM_LOWATER, fs_usage};
-    use super::super::metadata::MIN_MDA_SECTORS;
-    use super::super::tests::{loopbacked, real};
-    use super::super::tests::tempdir::TempDir;
 
     use super::*;
